# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Seatmap SaaS API
  version: "1.0.0"
  summary: Multi-tenant seating inventory, holds, allocations, ticketing and check-in.
  description: |
    The API owns venue seating inventory. It does **not** own carts, payments, tax or refunds —
    those stay in the tenant's own storefront (see `docs/adr/0001-saas-woocommerce-boundary.md`).

    ## Authentication
    Three separate schemes, deliberately not interchangeable:

    | Audience | Scheme | Notes |
    | --- | --- | --- |
    | Tenant panel / management API | `bearerAuth` (personal access token) | Scoped to the caller's tenant and role. |
    | Server-to-server (WordPress plugin) | `apiKeyAuth` + `hmacSignature` | All four `X-Seatmap-*` headers required. |
    | Embed widget (browser) | none | Public event id only; never carries a secret. |
    | Check-in device | `bearerAuth` (device token) | Scoped to named events. |

    ## HMAC signing
    `signature = hex(hmac_sha256(secret, METHOD + "\n" + PATH + "\n" + timestamp + "\n" + nonce + "\n" + hex(sha256(raw_body))))`

    `secret` is the value returned once when the key was created. The path is part of the signed
    string, so a captured signature cannot be redirected to another endpoint.

    Timestamps must be within ±300 s of server time; nonces may not repeat inside that window.
    Both conditions are enforced, so an accurate clock on the calling server is a requirement,
    not a nicety.

    ## Idempotency
...

    ## Errors
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /team/members/{member}:
    parameters:
      - name: member
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateMember
      tags: [Team]
      summary: Change someone's role, or suspend and restore them
      description: |
        Two rules are enforced here and both exist because of a way an account can become unusable:

        - **Nobody changes their own membership.** Without it the permission system is advice —
          anyone with `team.manage` could hand themselves everything.
        - **An account keeps at least one owner who is not suspended.** Otherwise nobody can grant
          anything ever again, and the only way out is a support ticket and a database console.

        Suspension rather than deletion, so somebody who left keeps their name on what they did.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, description: A built-in role key or one this organiser invented. }
                suspended: { type: boolean }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Member" }
        "403":
          description: "`cannot_change_own_role`, or the caller lacks `team.manage`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: "`last_owner`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /team/invitations/{invitation}:
    parameters:
      - name: invitation
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: revokeInvitation
      tags: [Team]
      summary: Cancel an invitation
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: boolean }
...
  /roles/{role}:
    parameters:
      - name: role
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateRole
      tags: [Team]
      summary: Rename a role or change what it holds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 80 }
                permissions: { type: array, items: { type: string } }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Role" }
    delete:
      operationId: deleteRole
      tags: [Team]
      summary: Delete a role nobody holds
      description: Deleting a role somebody holds would leave them with nothing, silently.
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: boolean }
        "409":
          description: "`role_in_use`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Seatmap-Key
      description: Public key id of an API key. Never the secret.
    hmacSignature:
      type: apiKey
      in: header
      name: X-Seatmap-Signature
      description: |
        Requires `X-Seatmap-Timestamp` (unix seconds, ±300 s) and `X-Seatmap-Nonce`
        (unique within the window) alongside.
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier.
              examples: [seat_unavailable, capacity_unavailable, hold_expired, invalid_signature, tenant_limit_reached]
            message: { type: string }
            details:
              type: object
              additionalProperties: true
            request_id: { type: string }
...
    Member:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        name: { type: [string, "null"] }
        email: { type: [string, "null"] }
        role: { type: string }
        role_name: { type: string, description: Translated for built-in roles; the organiser's own wording otherwise. }
        suspended: { type: boolean }
        last_seen_at: { type: [string, "null"], format: date-time }
...
    Role:
      type: object
      properties:
        id:
          type: [string, "null"]
          format: uuid
          description: Null for a built-in role, which is code rather than a row.
        key: { type: string }
        name: { type: string }
        description: { type: [string, "null"] }
        built_in: { type: boolean }
        permissions:
          type: array
          items: { type: string }
...